Added old button below and new btn above and added all the axios code in an external use-axios js in src/hooks.
