+++ src/pages/components/CertificateActionsButtons.js
@@ -1,17 +1,20 @@
 import React, { useState } from 'react'
-import axios from 'axios'
 import styled, { keyframes } from 'styled-components'
+
+import useAxios from '../../hooks/use-axios'
 
 import { Grid, Button } from '@digicatapult/ui-component-library'
 
 export default function CertificateActionsButtons() {
   // START
-  const [data, setData] = useState(null)
-  const [error, setError] = useState('')
-  const [loading, setLoading] = useState(false)
+  // const [data, setData] = useState(null)
+  // const [error, setError] = useState('')
+  // const [loading, setLoading] = useState(false)
   // END
 
   const [isWaitingVal, setIsWaitingVal] = useState(false)
+
+  const { data, error, loading, callApiFn } = useAxios()
 
   const handleClickSaveDraft = (e) => {
     e.preventDefault()
@@ -29,37 +32,41 @@ export default function CertificateActionsButtons() {
     }, 2000)
   }
 
+  // const { callApiFn } = useHttp()
+
   // START
   // check w/
   // u=http://localhost:8000/v1/certificate ; len=$(curl -s $u | jq length) ; curl -s $u | jq .[$((r - 1))]
-  const handleClick = () => {
-    setLoading(true)
-    const url = 'http://localhost:8000/v1/certificate'
-    const bodyObj = {
-      energy_consumed_wh: 2000000,
-      production_start_time: '2024-01-25T10:00:00.000Z',
-      production_end_time: '2024-01-25T20:00:00.000Z',
-      regulator: 'Reginald',
-      energy_owner: 'Emma',
-      hydrogen_quantity_wh: 2000000,
-    }
-    const defaultHeaders = { 'content-type': 'application/json' }
-
-    axios
-      .post(url, bodyObj, { headers: defaultHeaders })
-      .then(async (res) => {
-        // Eight second artificial delay
-        const wait = (ms) => new Promise((res) => setTimeout(res, ms))
-        await wait(8000)
-
-        setData(res.data)
-      })
-      .catch((err) => {
-        setError(err)
-        setLoading(false)
-      })
-      .finally(() => setLoading(false))
-  }
+  // const handleClick = () => {
+  //   setLoading(true)
+  //   const url = 'http://localhost:8000/v1/certificate'
+  //   const bodyObj = {
+  //     energy_consumed_wh: 2000000,
+  //     production_start_time: '2024-01-25T10:00:00.000Z',
+  //     production_end_time: '2024-01-25T20:00:00.000Z',
+  //     regulator: 'Reginald',
+  //     energy_owner: 'Emma',
+  //     hydrogen_quantity_wh: 2000000,
+  //   }
+  //   const defaultHeaders = { 'content-type': 'application/json' }
+  //
+  //   axios
+  //     .post(url, bodyObj, { headers: defaultHeaders })
+  //     .then(async (res) => {
+  //       // Eight second artificial delay
+  //       const wait = (ms) => new Promise((res) => setTimeout(res, ms))
+  //       await wait(8000)
+  //
+  //       setData(res.data)
+  //     })
+  //     .catch((err) => {
+  //       setError(err)
+  //       setLoading(false)
+  //     })
+  //     .finally(() => setLoading(false))
+  //
+  //   callApiFn({ url, method: 'POST', body: bodyObj, headers: defaultHeaders })
+  // }
   // END
 
   return (
@@ -92,25 +99,32 @@ export default function CertificateActionsButtons() {
           </Grid.Panel>
           <Grid.Panel area="div-double">
             <LargeButton
+              onClick={() => callApiFn()}
+              type="button"
+              disabled={loading}
+              variant="roundedPronounced"
+            >
+              {loading == false && data == null && <Span>Submit</Span>}
+              {loading == false && data != null && <Span>Submitted</Span>}
+              {loading && <AnimatedSpan>...</AnimatedSpan>}
+              {error && <Span>Error</Span>}
+            </LargeButton>
+            <br />
+            <hr />
+            {data && (
+              <div style={{ width: '266px' }}>{JSON.stringify(data)}</div>
+            )}
+            <br />
+            <hr />
+
+            <LargeButton
               variant="roundedPronounced"
               onClick={handleClickSubmit}
             >
               Submit
               {isWaitingVal && <span>...</span>}
             </LargeButton>
-            <br />
-            <hr />
-            <LargeButton
-              onClick={handleClick}
-              disabled={loading}
-              variant="roundedPronounced"
-              type="button"
-            >
-              {loading == false && data == null && <span>Submit</span>}
-              {loading == false && data != null && <span>Submitted</span>}
-              {loading && <AnimatedSpan>...</AnimatedSpan>}
-              {error && <div>ERROR: {JSON.stringify(error)}</div>}
-            </LargeButton>
+
             {/*
             <button type="button" disabled={loading} onClick={handleClick}>
               {loading == false && data == null && <span>Submit</span>}
@@ -118,9 +132,6 @@ export default function CertificateActionsButtons() {
               {loading && <span>...</span>}
             </button>
             */}
-            {data && (
-              <div style={{ width: '266px' }}>{JSON.stringify(data)}</div>
-            )}
           </Grid.Panel>
         </Grid>
       </PaddedDiv>
@@ -181,3 +192,5 @@ const AnimatedSpan = styled.span`
   margin: 0 auto;
   animation: ${RevealAnimation} 1s steps(4, end) infinite;
 `
+
+const Span = styled.span``
